Remove imu related joints and links from urdf
The urdf now only gives the TFs between the base_link and the other
parts of the robot. A dynamic publisher will eventually publish the TF
between the base_link and base_stabilized (given by the roll and pitch
values from the IMU). There will then be a static publisher publishing
the TF between the base_stabilized and the base_footprint.