Remove old laser_to_cloud converter node
This node converted the laser scan into a point cloud. The output point cloud
was in the map_world_frame. Since we are now using octomapping, we want
to pass the point cloud into the octomapping node in the laser scan
frame and the octomap node will transform it to the map_world_frame. The
new laser_to_cloud converter node will keep the converted point cloud
in the original laser frame.