Rename laser_assembler_implementation pkg
We are no longer using the laser_assembler pkg, so I'm changing the name
of my package to reflect it's use. I am going give it a broader name so
that it's name does not require a change if we change the nodes/pkgs
that we use in the future.