Change base_footprint height to reflect robot setup
The base_footprint is now accurately offset from the base_link
so that the 3d map is built directly on top of the 2D map. Note.
this offset is only relevant for the current setup of the unit
sitting on an milk crate, on the movable wheelchair base. This
offset will need to be changed when a new robot setup is used.