Change initial_tf_pub node to not account for yaw
The corrected imu frame now only accounts for roll and pitch. This means
that the yaw value is not offset, so the raw value from the imu is used.
This change means that the heading of the inertial and map frame is no
longer aligned.

This was changed since the static yaw offset that was previously being
used was causing issues due to the large drift and inaccurate yaw
readings from the imu.